+++ src/csp_doctor/__init__.py
@@ -1,4 +1,4 @@
 """CSP Doctor package."""
 
 __all__ = ["__version__"]
-__version__ = "0.1.17"
+__version__ = "0.1.18"